Config-gargoyle.in: enhance configuration defaults
Where it makes sense the Gargoyle kconfig can be used to set
symbol and package defaults for the directly supported targets
to reduce reliance on, and maintenance of, the individual target
profile config files.

Enhancements include:
- include some additional general config symbol settings for all
  builds
- add additional menu items for package groups such as target
  specific ethernet drivers, optional ethernet drivers, optional
  wifi drivers and optional utilities (including opkg)
- fine tune some symbol and/or package inclusions for specific targets

Limitations of the kconfig system mean that symbols defaulting to
enabled (i.e. "=m" or "=y") that need be disabled (i.e. set "=n")
can't be incorporated in the Gargoyle kconfig and must be unset in
the target profile config(s) instead.

Signed-off-by: Andrew MacIntyre <andymac@pcug.org.au>